+++ src/features/agency/agencyComponents/AgencyContent.jsx
@@ -1,13 +1,13 @@
 import styles from "./AgencyContent.module.css";
 import { useEffect } from "react";
-import Calendar from "../../package/calendar/Calendar";
-import { Hstack, Vstack } from "../../package/layout";
-import RoundBox from "../../package/RoundBox";
-import useLinkUpStore from "../../shared/store/store";
+import Calendar from "../../../package/calendar/Calendar";
+import { Hstack, Vstack } from "../../../package/layout";
+import RoundBox from "../../../package/RoundBox";
+import useLinkUpStore from "../../../shared/store/store";
 import AgencyModal from "./AgencyModal";
 import AgencySidebar from "./AgencySidebar";
-import { dummyEventArray } from "../../shared/store/dummyThepott";
-import FlexOneContainer from "../../package/flexOneContainer/FlexOneContainer";
+import { dummyEventArray } from "../../../shared/store/dummyThepott";
+import FlexOneContainer from "../../../package/flexOneContainer/FlexOneContainer";
 
 const AgencyContent = () => {
     const eventArray = useLinkUpStore((state) => state.eventArray);
