+++ src/pages/AgencyPage.jsx
@@ -1,5 +1,5 @@
-import AgencyContent from "../features/agency/AgencyContent";
-import AgencySkeleton from "../features/agency/AgencySkeleton";
+import AgencyContent from "../features/agency/agencyComponents/AgencyContent.jsx";
+import AgencySkeleton from "../features/agency/agencyComponents/AgencySkeleton.jsx";
 
 const AgencyPage = () => {
     const isLoading = false;
